Fixed serious bug: the register's CLEAR pin is now pulled high instead of low.
The CLEAR pin of the  74HC595 was accidentally connected to ground,
pulling it low and thereby active all the time, preventing the register
from working normally. It is now pulled to high and thereby inactive,
allowing normal operation.

The board now bears the version 1.1. The board has been updated to use the
parts from the new sub-libraries. The last throughole resister has been
replaced with an 1206 surface mount version. The lower right mounting has
been moved to the left of the connector and can now actually be used even
with a locking connector attached.

All of the writing is now on the appropriate layer.